template for socket and change interface logic
