+++ CMakeLists.txt
@@ -10,7 +10,6 @@ add_executable(${PROJECT_NAME}
     ${SRC_DIR}/interface.cpp
     ${SRC_DIR}/route.cpp
     ${SRC_DIR}/rip_packet.cpp
-    ${SRC_DIR}/rip_socket.cpp
     ${SRC_DIR}/system_netlink.cpp
     ${SRC_DIR}/rip.cpp
 )
