+++ rip/common.hpp
@@ -3,11 +3,16 @@
 
 #include <asio.hpp>
 
+#include "rip_protocol.hpp"
+
 #define DEBUG
 
-struct rip_socket;
+template<typename T, uint32_t port>
+class rip_socket;
+
+class rip_packet;
 
 extern asio::io_context service;
-extern rip_socket rip_sock;
+extern rip_socket<rip_packet, RIP_PORT> rip_sock;
 
 #endif /* COMMON_HPP */
