+++ rip/interface.cpp
@@ -9,6 +9,7 @@
 #include "common.hpp"
 #include "interface.hpp"
 
+/* TODO: change shared to raw pointer */
 std::list<std::shared_ptr<interface>> interface_list;
 
 #ifdef DEBUG
@@ -23,22 +24,21 @@ void interface_show()
 }
 #endif
 
-std::shared_ptr<interface> interface_get_shared_by_name(const char *iface_name, bool create)
+interface *interface_get_by_name(const char *iface_name, bool create)
 {
-    std::shared_ptr<interface> iface_ptr = nullptr;
+    interface *iface_ptr = nullptr;
 
     if (!interface_list.empty()) {
         for ( std::shared_ptr<interface> iface : interface_list) {
-            std::shared_ptr<interface> inteface_get_by_name(const char *iface_name, bool create);
             if (iface->name.compare(iface_name) == 0)
-                iface_ptr = iface;
+                iface_ptr = iface.get();
         }
     }
 
     if (iface_ptr == nullptr & create) {
         std::shared_ptr<interface> iface = std::make_shared<interface>(interface(iface_name));
         interface_list.push_back(iface);
-        iface_ptr = iface;
+        iface_ptr = iface.get();
     }
 
     return iface_ptr;
@@ -68,15 +68,4 @@ interface *interface_get_by_index(uint32_t index)
     }
 
     return nullptr;
-}
-
-bool interface_join_mcast_group(rip_socket &sock)
-{
-    std::for_each(interface_list.begin(), interface_list.end(), [&sock](std::shared_ptr<interface> iface) {
-        if (!(iface->flags & interface_flag_passive) && !iface->address.is_unspecified()) {
-            sock.join_mcast_group(iface->address);
-        }
-    });
-
-    return true;
 }
