+++ rip/interface.hpp
@@ -7,9 +7,6 @@
 #include <unordered_map>
 #include <memory>
 
-#include "route.hpp"
-#include "rip_socket.hpp"
-
 #define interface_flag_running 0x0001
 #define interface_flag_passive 0x0100
 
@@ -27,8 +24,6 @@ struct interface {
 	asio::ip::address_v4 address;
 	uint8_t address_prefix;
 
-	std::unordered_map<int, std::shared_ptr<struct route>> route_htable;
-
 	interface(const char *iface_name) : name(iface_name) {
 		this->index = 0;
 
@@ -39,12 +34,9 @@ struct interface {
 };
 
 void interface_show();
-std::shared_ptr<interface> interface_get_shared_by_name(const char *iface_name, bool create);
-std::shared_ptr<interface> interface_get_shared_by_index(const char *iface_name, bool create);
+interface *interface_get_by_name(const char *iface_name, bool create);
 interface *interface_get_by_index(uint32_t index);
-bool interface_join_mcast_group(rip_socket &sock);
 
 /* for_each functions */
-// void interface_update
 
 #endif /* INTERFACE_HPP */
